requirements.txt
addict==2.4.0
aider-chat==0.86.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
ConfigArgParse==1.7.1
cssselect==1.3.0
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
espeakng-loader==0.2.4
executing==2.2.1
fastjsonschema==2.21.2
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.18.0
flake8==7.3.0
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grep-ast==0.9.0
groq==0.37.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kokoro==0.9.4
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.63
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.2.11
language-tags==1.2.0
litellm==1.75.0
loguru==0.7.3
lxml==6.0.2
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
misaki==0.9.4
mixpanel==4.10.1
mpmath==1.3.0
mslex==1.3.0
multidict==6.6.3
murmurhash==1.0.15
mypy_extensions==1.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.2
num2words==0.5.14
numpy==1.26.4
openai==1.99.1
orjson==3.11.5
oslex==0.1.3
packaging==24.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.5.1
posthog==6.4.1
preshed==3.0.12
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==7.5.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-file==3.0.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
segments==2.3.0
sgmllib3k==1.0.0
shtab==1.7.2
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
thinc==8.3.10
tiktoken==0.10.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.9.0
tree-sitter-yaml==0.7.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.3
Werkzeug==3.1.4
wrapt==2.0.1
yarl==1.20.1
zipp==3.23.0
